+++ api/src/main/java/com/homeaway/datapullclient/process/DataPullTask.java
@@ -227,6 +227,7 @@ public class DataPullTask implements Runnable {
                 .withHadoopJarStep(runExampleConfig);
 
         final Application spark = new Application().withName("Spark");
+        final Application hive = new Application().withName("Hive");
 
         final EMRProperties emrProperties = this.config.getEmrProperties();
         final int instanceCount = emrProperties.getInstanceCount();
@@ -282,17 +283,43 @@ public class DataPullTask implements Runnable {
                 .withClassification("emrfs-site")
                 .withProperties(emrfsProperties);
 
+        Map<String, String> sparkHiveProperties = emrProperties.getSparkHiveProperties().entrySet().stream()
+                .filter(keyVal -> !keyVal.getValue().isEmpty())
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        sparkHiveProperties.putAll(this.clusterProperties.getSparkHiveProperties());
+
+        Map<String, String> hiveProperties = emrProperties.getHiveProperties().entrySet().stream()
+                .filter(keyVal -> !keyVal.getValue().isEmpty())
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        hiveProperties.putAll(this.clusterProperties.getHiveProperties());
+
+        Configuration sparkHiveConfig = new Configuration()
+                .withClassification("spark-hive-site")
+                .withProperties(sparkHiveProperties);
+
+        Configuration hiveConfig = new Configuration()
+                .withClassification("hive-site")
+                .withProperties(hiveProperties);
+
         final RunJobFlowRequest request = new RunJobFlowRequest()
                 .withName(this.taskId)
                 .withReleaseLabel(Objects.toString(this.clusterProperties.getEmrReleaseVersion(), emrReleaseVersion))
                 .withSteps(customExampleStep)
-                .withApplications(spark)
+                .withApplications(spark, hive)
                 .withLogUri(logPath)
                 .withServiceRole(Objects.toString(this.clusterProperties.getEmrServiceRole(), serviceRole))
                 .withJobFlowRole(Objects.toString(this.clusterProperties.getInstanceProfile(), jobFlowRole))  //addAdditionalInfoEntry("maximizeResourceAllocation", "true")
                 .withVisibleToAllUsers(true)
                 .withTags(this.emrTags.values()).withConfigurations(new Configuration().withClassification("spark").withProperties(sparkProperties), myEmrfsConfig)
                 .withInstances(jobConfig);
+
+        if (!hiveProperties.isEmpty()) {
+            request.withConfigurations(hiveConfig);
+        }
+
+        if (!sparkHiveProperties.isEmpty()) {
+            request.withConfigurations(sparkHiveConfig);
+        }
 
         if (!emrSecurityConfiguration.isEmpty()) {
             request.withSecurityConfiguration(emrSecurityConfiguration);